Deprecated Translate document, added action group names
