+++ Apps.DeepL/Apps.DeepL.csproj
@@ -5,12 +5,12 @@
         <Nullable>enable</Nullable>
         <Product>DeepL</Product>
         <Description>The NMT service</Description>
-        <Version>1.2.0</Version>
+        <Version>1.2.1</Version>
         <AssemblyName>Apps.DeepL</AssemblyName>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Blackbird.Applications.SDK.Blueprints" Version="0.0.1-rc.9597" />
+        <PackageReference Include="Blackbird.Applications.SDK.Blueprints" Version="1.0.0" />
         <PackageReference Include="Blackbird.Applications.SDK.Extensions.FileManagement" Version="1.0.1" />
         <PackageReference Include="Blackbird.Applications.Sdk.Glossaries.Utils" Version="1.0.0-alpha2" />
         <PackageReference Include="Blackbird.Applications.Sdk.Utils" Version="1.0.26" />
